+++ CHANGELOG.md
@@ -2,8 +2,7 @@
 ## 2.22 branch
 ### 2.22.0
 * a massive improvement of pages and posts templates;
-* fixed all admin templates for Bootstrap 4;
-* fixed `login` layout for Bootstrap 4;
+* fixed all templates and layouts for Bootstrap 4;
 * fixed `MenuBuilderHelper` for Bootstrap 4;
 * added the `post-preview` view element;
 * `MenuHelper` class returns menus as arrays, without transforming them into